usb/pcixhci: Move RP1 platform discovery to arch-specific override
