+++ arch/aarch64-raspi/usb/pcixhci/make.opts
@@ -1,0 +1,2 @@
+USER_CPPFLAGS += -DPCIXHCI_PLATFORM_PROBE
+FILES += $(SRCDIR)/arch/aarch64-raspi/usb/pcixhci/pcixhci_platform
